Refactoring the repo
