+++ README.md
@@ -3,3 +3,9 @@
 This application is designed to generate the schema for given SQL tables as GO structs. 
 
 Please take a look at the example directory for a sample schema layout and the output of the application.
+
+## Installation
+
+```bash
+go install github.com/jacobbrewer1/goschema@latest
+```
