+++ cmd_migrate.go
@@ -1,4 +1,4 @@
-package main
+package goschema
 
 import (
 	"context"
