+++ pkg/generation/template_helpers_test.go
